Cleanup code after refactoring
